Ykmd 5.0 SCP11b/MSVC updates

+++ aes_cmac/aes.h
@@ -36,7 +36,6 @@
 #ifdef _WIN32
 #include <windows.h>
 #include <bcrypt.h>
-#include <ntstatus.h>
 #else
 #include <openssl/evp.h>
 #endif
@@ -57,7 +56,7 @@ typedef struct {
     BCRYPT_KEY_HANDLE hKeyECB;
     PBYTE pbKeyCBCObj;
     PBYTE pbKeyECBObj;
-    size_t cbKeyObj;
+    DWORD cbKeyObj;
 #else
     EVP_CIPHER_CTX *ctx;
     unsigned char key_algo;
